Added seller field to Partner model.
Currently a Partner belongs to one Seller, we might
move to a situation in the future where the relationship between Seller
and Partner is many to many. One Seller can have multiple fulfilment
Partners and a fulfilment Partner can work with multiple sellers.

We will change the data model accordingly in the future