Multiple fixes. Turns out that the SellerProduct
was not being saved at the AddProduct step in the Seller Registration
wizard. The model is being saved now. This commit also fixes issue #25
by checking to see whether the user has already added products or not.
If a user who has registered visits the seller registration page again
then they are taken to the last Thanks step. and when they click the
Go to main shop. they should be taken to the main shop, and another
welcome email will not be sent