Changed registration/forms.py from windows line ending
to unix. Also made one change in AddProduct.clean method
to give the shop handle a default value if it is not
present in order to check for it. This seems to be
causing issues on my development environment, but it was
working earlier. Testing to see if this works in beta