Temporary fix for issue #55 Popup a message
when the user clicks the back button warning them that
the should not go back to edit and they can edit later.
TODO
Permanent fix is to allow the user to move back and forth
in the wizard and change things and revalidate all their
actions. We will work on this later