Fixed issues with Get OTP being clicked multiple times
Earlier fix was waiting for server response but the user can
click multiple times before the server responds. Now it sets the form
state to clicked and doesn't allow further clicks till there is an error
Email OTP also was looking for the same email address that was saved
in the user. Have created a check in the JsonResponseView to only
look for emails with the current email address that are not the current
user