Introduced an error step to seller registration
in case a 500 error is ever thrown. The error happens in the
get_form method of the WizardView. When the key, which is the current
step is not present in get_forms_list OrderedDict. We are storing data
about the incident in log files and pickle files. Hopefully this will
be enough to analyze what caused the error