Small commit but big changes! CMS toolbar no longer
appears on all pages of the site. Have to thank this guy
https://github.com/django-cms/django-cms/issues/3146#issuecomment-135046036
for the fix. I feel like I used this same fix years ago but
couldn't for the life of me find it anywhere.

This fix also means that things like input boxes appearing black
on certain pages because of the toolbar injecting its CSS/JS
in the page no longer happens