Made a mistake of creating migrations to load
and create objects in the database like a default system
User, a system Seller, Sites etc. But this approach was problematic
because if an object is ever changed after that script is created,
then the script fails to take into account the changed Object

Migrations have been cleaned up, so if you are following this repo,
then at this point you will have to drop your existing databases
and start again. I will try to create some future commit that
will backup and read from database backups