Added django-rules library to handle model level
permissions to view media files. Media files are not stored in
MEDIA_ROOT but rather in PROTECTED_ROOT where they are handled by
the django-downloadview library. We next need to commit code to
ensure proper integration with nginx. So that django can handle
permissions on the protected files but nginx will handle the actual
content delivery