Lots of changes here. Added admin_volt django
admin template to style the admin site. Seller Admin can
now approve and reject the Seller in the Seller detail screen.
Some javascript validation checks to ensure that the uploaded
files are approved before allowing the Approval to happen or
else in the case of rejection some text needs to be entered in
the Rejection Reasons text box.