Actually this should be two separate commits
but i'm being lazy. The context_processors and
templatetags introduce a template tag library called site_urls
which has a template tag called site_url which does a reverse
lookup on a viewname for a given site. the {% url %} only provides
the lookup under the current request context and current website
This allows us to get the URL links for pages on either site - seller
and buyer